refactor: improve code logics for lesson 21 to 25
